+++ tests/test_ac1015/test_image.py
@@ -80,7 +80,7 @@ class TestCreateNewImage(unittest.TestCase):
     def test_new_image_def(self):
         rootdict = self.dwg.rootdict
         self.assertFalse('ACAD_IMAGE_DICT' in rootdict)
-        imagedef = self.dwg.add_image_def('mycat', 'mycat.jpg', size_in_pixel=(640, 360))
+        imagedef = self.dwg.add_image_def('mycat.jpg', size_in_pixel=(640, 360))
 
         # check internals image_def_owner -> ACAD_IMAGE_DICT
         image_dict_handle = rootdict['ACAD_IMAGE_DICT']
@@ -97,7 +97,7 @@ class TestCreateNewImage(unittest.TestCase):
 
     def test_new_image(self):
         msp = self.dwg.modelspace()
-        image_def = self.dwg.add_image_def('mycat', 'mycat.jpg', size_in_pixel=(640, 360))
+        image_def = self.dwg.add_image_def('mycat.jpg', size_in_pixel=(640, 360))
         image = msp.add_image(insert=(0, 0), size_in_units=(3.2, 1.8), image_def=image_def)
         self.assertEqual((0, 0, 0), image.dxf.insert)
         self.assertEqual((0.005, 0, 0), image.dxf.u_pixel)