+++ NEWS.rst
@@ -2,7 +2,7 @@
 News
 ====
 
-Version 0.7.5 - 2016-..
+Version 0.7.5 - 2016-04-03
 
   * NEW: Drawing.acad_release property - AutoCAD release number for the drawing DXF version like 'R12' or 'R2000'
   * NEW: support for PDFUNDERLAY, DWFUNDERLAY and DGNUNDERLAY entities